refactor: utils패키지의 클래스를 model 패키지로 이동, utils 패키지 제거
