+++ src/main/java/racingcargame/model/movementgenerator/Move.java
@@ -1,4 +1,4 @@
-package racingcargame.utils;
+package racingcargame.model.movementgenerator;
 
 public interface Move {
     int getMoveValue();
