+++ src/main/java/racingcargame/model/numbergenerator/NumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcargame.utils;
+package racingcargame.model.numbergenerator;
 
 public interface NumberGenerator {
     int generateNumber(final int boundaryNumber);
