+++ src/main/java/racingcargame/model/numbergenerator/RandomNumberGenerator.java
@@ -1,6 +1,7 @@
-package racingcargame.utils;
+package racingcargame.model.numbergenerator;
 
 import java.util.Random;
+import racingcargame.model.numbergenerator.NumberGenerator;
 
 public class RandomNumberGenerator implements NumberGenerator {
     private final Random random = new Random();
